requirements.txt
requests
python-dotenv
flask
openai

langchain
langchain-community
langchain-core

google-generativeai
langchain-google-genai
langchain-huggingface
langchain-ollama
sentence-transformers

chromadb

markdown
unstructured
networkx